requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-phonenumber-field==6.0.0
gunicorn==20.1.0
phonenumberslite==8.12.41
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
